Make next.config.js editable, move default config part into uneditable file, import defaults into main config
