+++ src/nextjs/assets/next.config.js
@@ -1,36 +1,8 @@
-const jsonConfig = require('./next.config.json')
-delete jsonConfig['//']
-
-/**
- * Picks env variables with keys starting with `term`.
- * @param {string} [term] - A starting term of a key to pick.
- * @returns {Record<string, string>}
- */
-function pickEnvironmentByKey(term) {
-  return Object.entries(process.env).reduce(
-    (result, [key, value]) => {
-      if (key.startsWith(term)) {
-        result[key] = value
-      }
-
-      return result
-    },
-
-    {},
-  )
-}
+const defaultConfig = require('./next.config.defaults')
 
 /**
  * @type {import('next').NextConfig}
  */
-const config = {
-  ...jsonConfig,
-  serverRuntimeConfig: pickEnvironmentByKey('NEXT_SERVER_'),
+module.exports = {
+  ...defaultConfig,
 }
-
-config.publicRuntimeConfig = {
-  ...(config.publicRuntimeConfig || {}),
-  ...pickEnvironmentByKey('NEXT_PUBLIC_'),
-}
-
-module.exports = config
